API Reference

HelpPlots

HelpPlots.assert_type_and_record_argument — Function

assert_type_and_record_argument(dispatch::AbstractString, type::Type, description::AbstractString; kw...)

Checks that exactly one keyword argument is provided and that its value is a subtype of the specified type.

HelpPlots.help_plot — Function

help_plot(args...; kw...)

Prints plotting arguments for objects

Call help_plot(...) just like you would call plot(...).

Returns the plot.

HelpPlots.help_plot! — Function

help_plot!(args...; kw...)

Prints plotting arguments for objects

Call help_plot!(...) just like you would call plot!(...).

Returns the plot.

HelpPlots.recipe_dispatch — Function

recipe_dispatch(args...)

Function used to return a string describing the recipe, starting from the arguments used for multiple dispatching of the recipie

recipe_dispatch(arg::Any)

Returns string representation of each of the arguments in the recipie. Packages using HelpPlots can specialize this for their specific types.